+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/template
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.12.2
+
 ## 0.0.18
 
 ### Patch Changes